refactor(task): migrate from client component to server component
deleted obsolete files and dir. readded package lock as tracking dependancy

supabase error auth. works but need to fix
