+++ .vscode/settings.json
@@ -1,0 +1,16 @@
+{
+	"cSpell.words": [
+		"daygrid",
+		"exceljs",
+		"fullcalendar",
+		"icam",
+		"kyziq",
+		"maintainee",
+		"nextjs",
+		"nextui",
+		"supabase",
+		"tailwindcss",
+		"trpc",
+		"YYMMDDHH"
+	]
+}
